Moved separate apis into their own subfolder and split up api functions into separate files; fixed index files so all apis can be accessed from a common `apis` object.
