+++ app/lib/api/apis/confirmation/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as post } from './post';
