+++ app/lib/api/apis/index.js
@@ -1,17 +1,19 @@
 // @flow
 
-import NotificationsApi from './notifications';
-import UsersApi from './users';
-import TopicsApi from './topics';
-import TokenApi from './token';
-import PasswordApi from './password';
-import ConfirmationApi from './confirmation';
+import * as confirmation from './confirmation';
+import * as notifications from './notifications';
+import * as password from './password';
+import * as token from './token';
+import * as topics from './topics';
+import * as users from './users';
 
-export {
-  NotificationsApi,
-  UsersApi,
-  TopicsApi,
-  TokenApi,
-  PasswordApi,
-  ConfirmationApi,
+const api = {
+  confirmation,
+  notifications,
+  password,
+  token,
+  topics,
+  users,
 };
+
+export default api;
