+++ app/lib/api/apis/notifications/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as getAll } from './getAll';
