+++ app/lib/api/apis/password/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as post } from './post';
