+++ app/lib/api/apis/token/index.js
@@ -1,0 +1,4 @@
+// @flow
+
+export { default as destroy } from './destroy';
+export { default as post } from './post';
