+++ app/lib/api/apis/token/post.js
@@ -3,10 +3,9 @@
  * API docs: https://openwebslides.github.io/documentation/#token-api
  */
 
-import * as m from '../model';
-import ApiRequest from '../ApiRequest';
-
-import { TOKEN_ENDPOINT } from './helpers/endpoints';
+import * as m from '../../model';
+import ApiRequest from '../../ApiRequest';
+import { TOKEN_ENDPOINT } from '../helpers/endpoints';
 
 const post = (email: string, password: string): Promise<m.ApiResponseData> => {
   const request = new ApiRequest();
@@ -29,20 +28,4 @@ const post = (email: string, password: string): Promise<m.ApiResponseData> => {
   return request.execute();
 };
 
-const destroy = (token: m.Token): Promise<m.ApiResponseData> => {
-  const request = new ApiRequest();
-
-  request
-    .setEndpoint(TOKEN_ENDPOINT)
-    .setMethod(m.methodTypes.DELETE)
-    .setToken(token);
-
-  return request.execute();
-};
-
-const TokenApi = {
-  post,
-  destroy,
-};
-
-export default TokenApi;
+export default post;
