+++ app/lib/api/apis/topics/getContent.js
@@ -1,0 +1,25 @@
+// @flow
+/**
+ * API docs: #TODO
+ */
+
+import type { Identifier } from 'types/model';
+
+import * as m from '../../model';
+import ApiRequest from '../../ApiRequest';
+import { TOPICS_ENDPOINT, TOPICS_CONTENT_ENDPOINT } from '../helpers/endpoints';
+
+const getContent = (topicId: Identifier, token: string): Promise<m.ApiResponseData> => {
+  const request = new ApiRequest();
+
+  request
+    .setEndpoint(TOPICS_ENDPOINT)
+    .setResource(topicId)
+    .setSubEndpoint(TOPICS_CONTENT_ENDPOINT)
+    .setMethod(m.methodTypes.GET)
+    .setToken(token);
+
+  return request.execute();
+};
+
+export default getContent;
