+++ app/lib/api/apis/topics/index.js
@@ -1,0 +1,8 @@
+// @flow
+
+export { default as destroy } from './destroy';
+export { default as get } from './get';
+export { default as getAllByUserId } from './getAllByUserId';
+export { default as getContent } from './getContent';
+export { default as patchContent } from './patchContent';
+export { default as post } from './post';
