+++ app/lib/api/apis/users/index.js
@@ -1,0 +1,4 @@
+// @flow
+
+export { default as get } from './get';
+export { default as post } from './post';
