+++ app/lib/api/apis/users/post.js
@@ -3,24 +3,9 @@
  * API docs: https://openwebslides.github.io/documentation/#users-api
  */
 
-import type { Identifier } from 'types/model';
-
-import * as m from '../model';
-import ApiRequest from '../ApiRequest';
-
-import { USERS_ENDPOINT } from './helpers/endpoints';
-
-const get = (id: Identifier, token: ?string): Promise<m.ApiResponseData> => {
-  const request = new ApiRequest();
-
-  request
-    .setEndpoint(USERS_ENDPOINT)
-    .setMethod(m.methodTypes.GET)
-    .setResource(id)
-    .setToken(token);
-
-  return request.execute();
-};
+import * as m from '../../model';
+import ApiRequest from '../../ApiRequest';
+import { USERS_ENDPOINT } from '../helpers/endpoints';
 
 const post = (
   email: string,
@@ -52,9 +37,4 @@ const post = (
   return request.execute();
 };
 
-const UsersApi = {
-  get,
-  post,
-};
-
-export default UsersApi;
+export default post;
