+++ app/lib/api/index.js
@@ -1,27 +1,11 @@
 // @flow
 
-import {
-  NotificationsApi,
-  UsersApi,
-  TopicsApi,
-  TokenApi,
-  PasswordApi,
-  ConfirmationApi,
-} from './apis';
 import type {
   ApiResponseData,
   Token,
 } from './model';
 
-export {
-  NotificationsApi,
-  UsersApi,
-  TopicsApi,
-  TokenApi,
-  PasswordApi,
-  ConfirmationApi,
-};
-
+export { default } from './apis';
 export type {
   ApiResponseData,
   Token,
