+++ app/modules/authentication/saga/api/confirmation.js
@@ -2,7 +2,7 @@
 
 import { call } from 'redux-saga/effects';
 
-import { ConfirmationApi } from 'lib/api';
+import apis from 'lib/api';
 
 import * as t from '../../actionTypes';
 
@@ -11,7 +11,7 @@ export const apiPostConfirmationSaga = function* (
 ): Generator<*, *, *> {
   try {
     const { email } = action.payload;
-    yield call(ConfirmationApi.post, email);
+    yield call(apis.confirmation.post, email);
   }
   catch (error) {
     // TODO: fix saga error handling
