+++ app/modules/authentication/saga/api/password.js
@@ -2,14 +2,14 @@
 
 import { call } from 'redux-saga/effects';
 
-import { PasswordApi } from 'lib/api';
+import apis from 'lib/api';
 
 import * as t from '../../actionTypes';
 
 export const apiPostPasswordSaga = function* (action: t.ApiPostPasswordAction): Generator<*, *, *> {
   try {
     const { email } = action.payload;
-    yield call(PasswordApi.post, email);
+    yield call(apis.password.post, email);
   }
   catch (error) {
     // TODO
